RSTRING_LEN() should return a long like in MRI

+++ lib/cext/ruby.h
@@ -553,7 +553,7 @@ enum ruby_coderange_type {
 
 char *RSTRING_PTR(VALUE string);
 int rb_str_len(VALUE string);
-#define RSTRING_LEN(str) rb_str_len(str)
+#define RSTRING_LEN(str) (long)rb_str_len(str)
 #define RSTRING_LENINT(str) rb_str_len(str)
 VALUE rb_intern_str(VALUE string);
 VALUE rb_str_new(const char *string, long length);
